+++ csfieldguide/config/__init__.py
@@ -1,3 +1,3 @@
 """Module for Django system configuration."""
 
-__version__ = "3.6.0"
+__version__ = "3.7.0"